# api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Dim active API", "version": "1.0.0" },
  "servers": [{ "url": "/api/v1" }],
  "paths": {
...
    "/library": {
      "get": { "operationId": "listLibraries", "security": [{ "token": [] }], "responses": { "200": { "description": "Libraries", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Library" } } } } }, "401": { "$ref": "#/components/responses/Error" } } },
      "post": { "operationId": "createLibrary", "security": [{ "token": [] }], "requestBody": { "description": "Library", "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateLibrary" } } } }, "responses": { "200": { "description": "Created library", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Library" } } } }, "400": { "$ref": "#/components/responses/Error" } } }
    },
    "/library/{id}": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "operationId": "getLibrary", "security": [{ "token": [] }], "responses": { "200": { "description": "Library", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Library" } } } }, "404": { "$ref": "#/components/responses/Error" } } }, "delete": { "operationId": "deleteLibrary", "security": [{ "token": [] }], "responses": { "204": { "description": "Deleted" }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/library/{id}/media": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "operationId": "getLibraryMedia", "security": [{ "token": [] }], "responses": { "200": { "description": "Media grouped under the library name", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": { "type": "array", "items": { "$ref": "#/components/schemas/Media" } } } } } } } } },
    "/library/{id}/unmatched": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "operationId": "getUnmatched", "security": [{ "token": [] }], "responses": { "200": { "description": "Unmatched files", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UnmatchedFiles" } } } } } } },
...
  },
  "components": {
    "securitySchemes": { "token": { "type": "apiKey", "in": "header", "name": "Authorization" }, "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "dim_session" } },
    "parameters": { "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }, "Gid": { "name": "gid", "in": "path", "required": true, "schema": { "type": "string" } } },
...
    "responses": { "Login": { "description": "Token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } }, "Register": { "description": "Created account and token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterResponse" } } } }, "Error": { "description": "Structured error", "headers": { "x-request-id": { "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } } } } },
    "schemas": {
      "ApiErrorEnvelope": { "type": "object", "required": ["error", "request_id"], "properties": { "error": { "$ref": "#/components/schemas/ApiError" }, "request_id": { "type": "string" } } },
      "ApiError": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "details": { "type": "object", "additionalProperties": true } } },
...
      "LoginResponse": { "type": "object", "required": ["token"], "properties": { "token": { "type": "string" } } },
      "RegisterResponse": { "type": "object", "required": ["token", "username"], "properties": { "token": { "type": "string" }, "username": { "type": "string" } } },
...
      "CreateLibrary": { "type": "object", "required": ["name", "locations", "media_type"], "properties": { "name": { "type": "string" }, "locations": { "type": "array", "items": { "type": "string" } }, "media_type": { "type": "string", "enum": ["movie", "tv"] } } },
      "Library": { "type": "object", "required": ["id", "name", "locations", "media_type"], "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "locations": { "type": "array", "items": { "type": "string" } }, "media_type": { "type": "string" }, "scan_status": { "type": "string", "enum": ["scanning", "complete", "failed"] } }, "additionalProperties": true },
      "Chapters": { "type": "object", "required": ["credits"], "properties": { "credits": { "type": "integer" } } },
      "Media": { "type": "object", "required": ["id", "name", "library_id", "media_type", "genres", "duration", "progress"], "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "library_id": { "type": "integer" }, "media_type": { "type": "string" }, "genres": { "type": "array", "items": { "type": "string" } }, "duration": { "type": "integer" }, "progress": { "type": "integer" }, "description": { "type": ["string", "null"] }, "added": { "type": ["string", "null"] }, "backdrop_path": { "type": ["string", "null"] }, "poster_path": { "type": ["string", "null"] }, "chapters": { "$ref": "#/components/schemas/Chapters" }, "episode": { "type": "integer" }, "next_episode_id": { "type": "integer" }, "play_btn_id": { "type": "integer" }, "prev_episode_id": { "type": "integer" }, "rating": { "type": "number" }, "season": { "type": "integer" }, "year": { "type": "integer" }, "tags": { "type": "object", "additionalProperties": true } }, "additionalProperties": true },
      "UnmatchedMediaFile": { "type": "object", "required": ["type"], "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "folder": { "type": "string" }, "duration": { "type": "integer" }, "file": { "type": "string" }, "files": { "type": "array", "items": { "$ref": "#/components/schemas/UnmatchedMediaFile" } }, "type": { "type": "string", "enum": ["directory", "file"] } } },
      "UnmatchedFiles": { "type": "object", "required": ["count", "files"], "properties": { "count": { "type": "integer" }, "files": { "type": "array", "items": { "$ref": "#/components/schemas/UnmatchedMediaFile" } } } },
...
    }
  }
}